requirements.txt
numpy
Pillow
dill
scipy
opencv-python
matplotlib
# open3d